+++ cloned_repos/test1/todo-app/tests/test_todo.py
@@ -1,11 +1,15 @@
-from todo import TodoManager
+#!/usr/bin/env python3
 
-def test_add():
+from todo_app.todo import TodoManager
+
+
+def test_add() -> None:
     manager = TodoManager()
     manager.add("Buy Milk")
     assert len(manager.todos) == 1
 
-def test_complete():
+
+def test_complete() -> None:
     manager = TodoManager()
     manager.add("Buy Milk")
     manager.complete("Buy Milk")